feat: final commit before merge MainPage

+++ src/main/java/com/github/moviereservationbe/web/DTO/movie/movieResponseDto.java
@@ -18,16 +18,16 @@ public class movieResponseDto {
     private String titleEnglish;
     private String poster;
     private Date releaseDate;
-    private Integer ticketSales;
-    private Integer scoreAvg;
-    //private Integer ageLimit;
-    //private Integer screenTime;
-    //private String country;
-    //private String director;
-    //private String genre;
+    private Double ticketSales;
+    private Double scoreAvg;
     private Integer dDay;
-    //private String status; //예매중, 현재상영중, 상영종료
-    //private String summary;
+//    private Integer ageLimit;
+//    private Integer screenTime;
+//    private String country;
+//    private String director;
+//    private String genre;
+//    private String status; //예매중, 현재상영중, 상영종료
+//    private String summary;
 
     public movieResponseDto(Movie movie) {
         this.movieId = movie.getMovieId();
@@ -39,7 +39,4 @@ public class movieResponseDto {
         this.poster = movie.getPoster();
         this.scoreAvg = movie.getScoreAvg();
     }
-
-
-
 }
